+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - new color style type 'intervals' [117](https://github.com/openkfw/Oscar/pull/120)
+- 'metadata.attributions' key in ./initial-data-load/data/{country}/GeoData.yml configuration file as a substite for 'tileAttributions' key on the top level that is now deprecated. You can find out more about these settings in [docs](./doc/data-structures/config-files.md#geoData.yml). [#146](https://github.com/openkfw/Oscar/pull/146/files)
+- 'metadata.attributions' key in ./initial-data-load/data/{country}/MapLayers.yml configuration file
 
 ### Changed
 
@@ -18,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - route '/api/staticLayers', use '/api/dataLayers' [111]()
 - 'COUNTRY' environment variable for initial-data-load, use 'DATASET' instead
-- 'tileDataUrl' and 'tileAttributions' keys on the top level in ./initial-data-load/data/{country}/GeoData.yml configuration file are now deprecated. Use 'apiUrl' key instead of 'tileDataUrl' key and 'metadata.description' key instead of 'tileAttributions' key. You can find out more about these settings in [docs](./doc/data-structures/config-files.md#geoData.yml).
+- 'tileDataUrl' and 'tileAttributions' keys on the top level in ./initial-data-load/data/{country}/GeoData.yml configuration file. Use 'apiUrl' key instead of 'tileDataUrl' key and 'metadata.attributions' key instead of 'tileAttributions' key. You can find out more about these settings in [docs](./doc/data-structures/config-files.md#geoData.yml). [#146](https://github.com/openkfw/Oscar/pull/146/files)
 
 ### Removed